Replace map store with global map state (#291)
* move button for minimized MapPosition into component

This means the button can no longer be a leaflet control.
But we can use svelte reactivity for switching between modes.

* convert MapPosition info fields to properties

Instead of registering events on the map itself, have the parent
map hand in all the info we want to display.

* make map state globally available

Removes the existing map_store and exports a globally available
set of map states instead. Components can now simply listen to
the map state using Svelte reactivity instead of registering their
own event handlers to the map.

* use Svelte runes for main app